+++ packages/dagonizer-tool-googlebooks/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @studnicky/dagonizer-tool-googlebooks
 
+## 1.0.1
+
+### Patch Changes
+
+- @studnicky/dagonizer-book-entities@1.0.1
+
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [fdaa32a]
+  - @studnicky/dagonizer@1.0.0
+  - @studnicky/dagonizer-book-entities@1.0.0
+
+## 0.30.1
+
+### Patch Changes
+
+- @studnicky/dagonizer-book-entities@0.30.1
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+  - @studnicky/dagonizer@1.0.0
+  - @studnicky/dagonizer-book-entities@1.0.0
+
 ## 0.30.0
 
 ### Patch Changes